Test for only members with roles being able to send pub messages

+++ src/basic_bot.py
@@ -87,10 +87,12 @@ async def itemMessage(message):
 
 async def generalMessage(message):
 	p = DictionaryReader()
+	roles = message.author.roles
 	command = message.content[1::].split(' ')[0].lower()
 	msg = p.commandReader(message.content[1::])
 	if msg != None:
-		if command in p.whisperCommands() or 'help' in message.content:
+		if len(roles) and (command in p.whisperCommands() or 'help' in message.content):
+			await client.send_message(message.channel, len(roles))
 			await client.send_message(message.author, msg)
 			await client.delete_message(message)
 		else:
